refactor: extract publishing contact form section
